@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 namespace FoodAPICore.Dtos
 {
-    public class FoodItemCreateDto
+    public class FoodCreateDto
     {
         public string Name { get; set; }
         public string Type { get; set; }
