@@ -1,11 +1,10 @@
 ﻿using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore;
 using Microsoft.Extensions.Logging;
 using System;
 using FoodAPICore.Entities;
 using Microsoft.Extensions.DependencyInjection;
 using FoodAPICore.Services;
-using Microsoft.AspNetCore.Identity;
+using Microsoft.Extensions.Hosting;
 
 namespace FoodAPICore
 {
@@ -13,7 +12,7 @@ namespace FoodAPICore
     {
         public static void Main(string[] args)
         {
-            var host = CreateWebHostBuilder(args).Build();
+            var host = CreateHostBuilder(args).Build();
 
             // Initializes db.
             using (var scope = host.Services.CreateScope())
@@ -35,8 +34,11 @@ namespace FoodAPICore
             host.Run();
         }
 
-        public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
-                .UseStartup<Startup>();
+        public static IHostBuilder CreateHostBuilder(string[] args) =>
+           Host.CreateDefaultBuilder(args)
+               .ConfigureWebHostDefaults(webBuilder =>
+               {
+                   webBuilder.UseStartup<Startup>();
+               });
     }
 }
